Added combined one pager
I combined the text from all the 4 one pagers into one document. I have used text from all 4 docs and spliced and reordered them in a way that made sense to me. However, note that my understanding of the social aspects is quite bad :-). So it may not make sense the way I arranged it. Please feel free to edit it as you seem fit.